Extract ConsoleEngineDiscoveryRequestListener
