+++ prototypes/javaspec-engine/src/main/java/info/javaspec/engine/JavaSpecEngine.java
@@ -9,14 +9,12 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ServiceLoader;
 
+//Discovers specs, turns them into something Jupiter can run, and runs them.
 public class JavaSpecEngine implements TestEngine {
   @Override
   public TestDescriptor discover(EngineDiscoveryRequest discoveryRequest, UniqueId engineId) {
     ServiceLoader<EngineDiscoveryRequestListener> loader = ServiceLoader.load(EngineDiscoveryRequestListener.class);
     loader.findFirst().ifPresent(x -> x.onDiscover(discoveryRequest));
-
-    System.out.printf("[JavaSpecEngine#discover]%n");
-    printDiscoveryRequest(discoveryRequest);
 
     EngineDescriptor engineDescriptor = new EngineDescriptor(engineId, "JavaSpec");
     discoveryRequest.getSelectorsByType(ClassSelector.class).stream()
@@ -31,65 +29,6 @@ public class JavaSpecEngine implements TestEngine {
       });
 
     return engineDescriptor;
-  }
-
-  //TODO KDK: Extract to ServiceLoader.  Put the interface in javaspec-api.
-  //Try this https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/ServiceLoader.html?is-external=true#load(java.lang.Class)
-  private void printDiscoveryRequest(EngineDiscoveryRequest discoveryRequest) {
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] ClasspathResourceSelector%n");
-    discoveryRequest.getSelectorsByType(ClasspathResourceSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getClasspathResourceName()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] ClasspathRootSelector%n");
-    discoveryRequest.getSelectorsByType(ClasspathRootSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getClasspathRoot()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] ClassSelector%n");
-    discoveryRequest.getSelectorsByType(ClassSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getClassName()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] DirectorySelector%n");
-    discoveryRequest.getSelectorsByType(DirectorySelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getRawPath()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] FileSelector%n");
-    discoveryRequest.getSelectorsByType(FileSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getRawPath()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] MethodSelector%n");
-    discoveryRequest.getSelectorsByType(MethodSelector.class)
-      .forEach(x -> System.out.printf("- %s#%s%n", x.getClassName(), x.getMethodName()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] PackageSelector%n");
-    discoveryRequest.getSelectorsByType(PackageSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getPackageName()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] UniqueIdSelector%n");
-    discoveryRequest.getSelectorsByType(UniqueIdSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getUniqueId()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] UriSelector%n");
-    discoveryRequest.getSelectorsByType(UriSelector.class)
-      .forEach(x -> System.out.printf("- %s%n", x.getUri().getRawPath()));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] ClassNameFilter%n");
-    discoveryRequest.getFiltersByType(ClassNameFilter.class)
-      .forEach(x -> System.out.printf("- %s%n", x));
-
-    System.out.println();
-    System.out.printf("[JavaSpecEngine#discover] PackageNameFilter%n");
-    discoveryRequest.getFiltersByType(PackageNameFilter.class)
-      .forEach(x -> System.out.printf("- %s%n", x));
   }
 
   private SpecClass makeDeclaringInstance(Class<SpecClass> specClass) {
